Move downloads teardown into makevers and update tests accordingly
Make sure to remove downloads after test_get_new

(cherry picked from commit eef2c990acf1d59849386fde7c653e3cfa6bf163)
