+++ test/test_get_new.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, print_function
 
 from ..pyautoupdate.launcher import Launcher
 from .pytest_skipif import needinternet
-from .pytest_makevers import fixture_update_dir
+from .pytest_makevers import fixture_update_dir,create_update_dir
 
 import os
 import sys
@@ -28,7 +28,7 @@ def create_zip(request):
 
 @pytest.mark.trylast
 @needinternet
-def test_check_get_new(create_zip, fixture_update_dir):
+def test_check_get_new(create_zip, fixture_update_dir,create_update_dir):
     """Test that gets new version from internet"""
     package=fixture_update_dir("0.0.1")
     launch = Launcher('filling up the boring replacements',
@@ -37,3 +37,16 @@ def test_check_get_new(create_zip, fixture_update_dir):
     with open(os.path.abspath("downloads/extradir/blah.py"), "r") as file_code:
         file_text=file_code.read()
     assert "new version" in file_text
+    assert os.path.isdir("downloads")
+
+@needinternet
+def test_check_get_invalid_archive(fixture_update_dir):
+    """Test that gets new version from internet"""
+    package=fixture_update_dir("0.0.1")
+    launch = Launcher('what file? hahahaha',
+                      r'http://rlee287.github.io/pyautoupdate/testing2/',
+                      newfiles="project.tar.gz")
+    launch._get_new()
+    assert os.path.isfile("project.tar.gz.dump")
+    assert not os.path.isdir("downloads")
+    os.remove("project.tar.gz.dump")
